+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx.hs
@@ -1,7 +1,9 @@
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableSuperClasses #-}
 
 -- | Transaction building and inspecting relies heavily on lenses (`microlens`). Therefore, some
 -- familiarity with those is necessary. However, you can probably go a long way by simply
@@ -39,6 +41,10 @@ module Cardano.Ledger.Api.Tx (
   module Cardano.Ledger.Api.Tx.AuxData,
   module Cardano.Ledger.Api.Tx.Wits,
 
+  -- * Any era
+  AnyEraTx (isValidTxG),
+  producedTxOuts,
+
   -- * Shelley onwards
   EraTx (Tx),
   mkBasicTx,
@@ -71,7 +77,7 @@ module Cardano.Ledger.Api.Tx (
 ) where
 
 import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsValid (..))
-import Cardano.Ledger.Api.Era (EraApi (..))
+import Cardano.Ledger.Api.Era
 import Cardano.Ledger.Api.Scripts.ExUnits (
   RedeemerReport,
   RedeemerReportWithLogs,
@@ -83,5 +89,45 @@ import Cardano.Ledger.Api.Tx.AuxData
 import Cardano.Ledger.Api.Tx.Body
 import Cardano.Ledger.Api.Tx.Cert
 import Cardano.Ledger.Api.Tx.Wits
+import Cardano.Ledger.Babbage.Collateral (mkCollateralTxIn)
 import Cardano.Ledger.Core (EraTx (..), binaryUpgradeTx, txIdTx)
+import Cardano.Ledger.State (UTxO (..), txouts)
 import Cardano.Ledger.Tools (calcMinFeeTx, estimateMinFeeTx, setMinFeeTx, setMinFeeTxUtxo)
+import Control.Monad (join)
+import qualified Data.Map as Map
+import Lens.Micro
+
+class (EraTx era, AnyEraTxBody era, AnyEraTxWits era, AnyEraTxAuxData era) => AnyEraTx era where
+  isValidTxG :: SimpleGetter (Tx era) (Maybe IsValid)
+  default isValidTxG :: AlonzoEraTx era => SimpleGetter (Tx era) (Maybe IsValid)
+  isValidTxG = isValidTxL . to Just
+
+instance AnyEraTx ShelleyEra where
+  isValidTxG = to (const Nothing)
+
+instance AnyEraTx AllegraEra where
+  isValidTxG = to (const Nothing)
+
+instance AnyEraTx MaryEra where
+  isValidTxG = to (const Nothing)
+
+instance AnyEraTx AlonzoEra
+
+instance AnyEraTx BabbageEra
+
+instance AnyEraTx ConwayEra
+
+instance AnyEraTx DijkstraEra
+
+-- | Construct all of the unspent outputs that will be produced by this transaction
+producedTxOuts :: AnyEraTx era => Tx era -> UTxO era
+producedTxOuts tx =
+  case tx ^. isValidTxG of
+    Just (IsValid False) ->
+      UTxO $
+        case join (txBody ^. collateralReturnTxBodyG) of
+          Nothing -> mempty
+          Just txOut -> Map.singleton (mkCollateralTxIn txBody) txOut
+    _ -> txouts txBody
+  where
+    txBody = tx ^. bodyTxL